% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bestrad_sequencing.R
\name{calculate_bestrad_sequencing}
\alias{calculate_bestrad_sequencing}
\title{BestRAD Sequencing Requirement Calculator}
\usage{
calculate_bestrad_sequencing(
  n_samples,
  c_val,
  enzyme,
  read_length_total = 300,
  lane_reads = 1e+09,
  quality_filter = 0.6,
  dup_removal = 0.75,
  desired_depth = 40
)
}
\arguments{
\item{n_samples}{Number of individuals in the project}

\item{c_val}{Genome size in C-value (required)}

\item{enzyme}{Name of the enzyme ("SbfI" or "PstI")}

\item{read_length_total}{Combined length of paired reads (default 300)}

\item{lane_reads}{Total reads per lane (default 1e9)}

\item{quality_filter}{Proportion of reads passing process_radtags (default 0.6)}

\item{dup_removal}{Proportion of reads retained after PCR duplicate removal (default 0.75)}

\item{desired_depth}{Target coverage depth per locus (default 40)}
}
\description{
BestRAD Sequencing Requirement Calculator
}
