% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genind_merge.R
\name{genind_merge}
\alias{genind_merge}
\title{Merge Multiple genind Objects by Common Loci}
\usage{
genind_merge(genind_input, filter_ntc = TRUE)
}
\arguments{
\item{genind_input}{A list of genind objects OR a character vector of
names of genind objects (e.g., c("dat1", "dat2")).}

\item{filter_ntc}{Logical; if TRUE, removes individuals with "ntc" in their
name (case-insensitive). Default is TRUE.}
}
\description{
Merge Multiple genind Objects by Common Loci
}
