% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gtscore_to_genind.R
\name{import_gtscore_to_genind}
\alias{import_gtscore_to_genind}
\title{Import GTscore to genind with Depth, Missingness Filtering, and Sample Reporting}
\usage{
import_gtscore_to_genind(
  genepop_path,
  allelereads_path,
  min_depth = 10,
  locus_max_missing = 0.2,
  sample_max_missing = 0.2,
  print_removed_inds = FALSE
)
}
\arguments{
\item{genepop_path}{Path to the Genepop file (works with .txt or .gen)}

\item{allelereads_path}{Path to the AlleleReads_singleSNPs.txt file}

\item{min_depth}{Integer; minimum total read depth (a1 + a2) to retain a genotype}

\item{locus_max_missing}{Numeric; max proportion of missing genotypes allowed for a locus (0-1)}

\item{sample_max_missing}{Numeric; max proportion of missing genotypes allowed for a sample (0-1)}

\item{print_removed_inds}{Logical; if TRUE, prints names of samples removed during filtering}
}
\description{
Import GTscore to genind with Depth, Missingness Filtering, and Sample Reporting
}
