% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_meer_sample_data.R
\name{import_meer_sample_data}
\alias{import_meer_sample_data}
\title{Fetch and Align Metadata for GTscore Samples}
\usage{
import_meer_sample_data(
  gen_obj,
  db_path = "Z:/meerLab_database_be.accdb",
  select_cols = NULL
)
}
\arguments{
\item{gen_obj}{The filtered genind object}

\item{db_path}{Path to the MS Access database (Default: MEER Lab Backend)}

\item{select_cols}{A character vector of column names. Defaults to everything().}
}
\description{
Fetch and Align Metadata for GTscore Samples
}
