% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gtscore_allele_depth_plots.R
\name{plot_all_gtscore_loci}
\alias{plot_all_gtscore_loci}
\title{Batch Plot Allele Balance from GTscore Output with Fixed Axes}
\usage{
plot_all_gtscore_loci(file_path, outdir, use_log_scale = FALSE, jitter = TRUE)
}
\arguments{
\item{file_path}{Full path to the AlleleReads_singleSNPs.txt file}

\item{outdir}{Directory where the locus plots will be saved}

\item{use_log_scale}{Logical; if TRUE, uses log10 scale for depths (default FALSE)}

\item{jitter}{Logical; if TRUE, adds small jitter to points (default TRUE)}
}
\description{
Batch Plot Allele Balance from GTscore Output with Fixed Axes
}
